Refactor theme logic into client provider
