+++ src/app/layout.tsx
@@ -1,8 +1,7 @@
-'use client'
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import { useEffect, useState } from "react";
 import "./globals.css";
+import ThemeProvider from "../components/theme-provider";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -20,35 +19,10 @@ export const metadata: Metadata = {
 };
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
-  const [theme, setTheme] = useState<'light' | 'dark'>('light');
-
-  useEffect(() => {
-    const stored = localStorage.getItem('theme') as 'light' | 'dark' | null;
-    if (stored) {
-      setTheme(stored);
-    } else if (window.matchMedia('(prefers-color-scheme: dark)').matches) {
-      setTheme('dark');
-    }
-  }, []);
-
-  useEffect(() => {
-    document.documentElement.classList.remove('light', 'dark');
-    document.documentElement.classList.add(theme);
-    localStorage.setItem('theme', theme);
-  }, [theme]);
-
   return (
-    <html lang="en" className={theme}>
+    <html lang="en">
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
-        <header className="fixed top-0 w-full p-4 flex justify-end z-10">
-          <button
-            onClick={() => setTheme(theme === 'light' ? 'dark' : 'light')}
-            className="px-3 py-1 rounded bg-[var(--accent)] text-[var(--bg)]"
-          >
-            {theme === 'light' ? 'Dark' : 'Light'} Mode
-          </button>
-        </header>
-        <main className="pt-16">{children}</main>
+        <ThemeProvider>{children}</ThemeProvider>
       </body>
     </html>
   );
